import to db from slack
